requirements.txt
mcpi
ini-parser
